tmpop: move evidence inside tmpop package
