+++ tmpop/tmpoptestcases/evidence.go
@@ -22,10 +22,10 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stratumn/go-indigocore/cs"
 	"github.com/stratumn/go-indigocore/cs/cstesting"
-	"github.com/stratumn/go-indigocore/cs/evidences"
 	"github.com/stratumn/go-indigocore/store"
 	"github.com/stratumn/go-indigocore/testutil"
 	"github.com/stratumn/go-indigocore/tmpop"
+	"github.com/stratumn/go-indigocore/tmpop/evidences"
 	"github.com/stratumn/go-indigocore/tmpop/tmpoptestcases/mocks"
 	"github.com/stratumn/go-indigocore/types"
 	"github.com/stratumn/merkle"
